Refactor qualifications type from vector<int> to vector<double> in Video class and comment out friend operator<< declarations in Chapter and Video headers

+++ Video.cpp
@@ -6,6 +6,12 @@ using namespace std;
 
 Video::Video(string type, string id, string title, string genre, string duration) : type(type), id(id), title(title), genre(genre), duration(duration) {}
 
+// ostream &operator<<(ostream &os, const Video &video)
+// {
+//     video.print(os); // ChatGPT
+//     return os;
+// }
+
 void Video::addScore(double score)
 {
     qualifications.push_back(score);
@@ -16,7 +22,7 @@ double Video::getAverage()
     if (qualifications.empty())
         return -1;
     double sum = 0.0;
-    for (int score : qualifications)
+    for (double score : qualifications)
     {
         sum += score;
     }
